feat(own): add plans and activity data

+++ app/own/migrations/0002_alter_ownfee_value_and_basemdr.py
@@ -10,11 +10,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="ownfee",
             name="value",
-            field=models.FloatField(),
+            field=models.DecimalField(decimal_places=10, max_digits=20),
         ),
         migrations.AlterField(
             model_name="ownfee",
             name="baseMdr",
-            field=models.FloatField(),
+            field=models.DecimalField(decimal_places=10, max_digits=20),
         ),
     ]
